requirements.txt
requests
beautifulsoup4
slack_sdk
python-dotenv